requirements.txt
discord.py
python-dotenv
supabase
sentence-transformers
torch
numpy
matplotlib
requests
pandas